Added comments/changes variables for clarity
